Most_Advanced_AI_Ever/requirements.txt
accelerate==0.21.0
aiofiles==22.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.4.0
audioread==3.0.1
babel==2.17.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
bidict==0.23.1
bitsandbytes==0.41.0
bleach==6.2.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
comtypes==1.4.12
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0

durationpy==0.10
edge-tts==7.0.2
entrypoints==0.4
executing==2.2.1
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.118.2
fastjsonschema==2.21.2
ffmpy==0.6.2
filelock==3.13.1
Flask==3.1.2
Flask-SocketIO==5.5.1
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.6.1
future==1.0.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
gradio==3.43.1
gradio_client==0.5.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.35.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.6.0
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
iso8601==2.1.0
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.context==6.0.1
jedi==0.19.2
Jinja2==3.1.4
jiter==0.11.0
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_fileid==0.9.3
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyboard==0.13.5
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==0.3.24
langchain-chroma==0.2.5
langchain-core==0.3.78
langchain-text-splitters==0.3.11
langsmith==0.3.45
lark==1.3.0
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multithreading==0.2.0
narwhals==2.7.0
nbclassic==1.3.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==6.5.7
notebook_shim==0.2.4
numba==0.62.1
numpy==1.26.1
oauthlib==3.3.1
ollama==0.1.8
onnxruntime==1.23.1
openai==0.27.8
openai-whisper==20250625
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.5
pdf2image==1.16.3
pdfminer.six==20250506
pdfplumber==0.11.7
Pillow==10.1.0
platformdirs==4.5.0
playsound==1.2.2
pooch==1.8.2
posthog==5.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==5.9.8
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pybase64==1.4.2
pycparser==2.23
pycryptodome==3.19.0
pydantic==2.12.0
pydantic_core==2.41.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyjokes==0.6.0
PyMsgBox==2.0.1
PyMuPDF==1.22.5
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.8.2
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.3
python-json-logger==4.0.0
python-multipart==0.0.20
python-socketio==5.13.0
python-vlc==3.0.21203
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
randfacts==0.22.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-learn==1.4.2
scipy==1.11.4
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
speechbrain==1.0.3
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
srt==3.5.3
stack-data==0.6.3
starlette==0.48.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.3.3
tinycss2==1.4.0
tokenizers==0.13.3

tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.30.2
typer==0.19.2
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.2.84
ultralytics-thop==2.0.17
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.37.0
vosk==0.3.45
watchfiles==1.1.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.9.0
Werkzeug==3.1.3
whisper==1.1.10
windows-curses==2.4.1
wolframalpha==5.1.3
word2number==1.1
wordtodigits==1.0.2
wsproto==1.2.0
xformers==0.0.23
xmltodict==1.0.2
y-py==0.6.2
yarl==1.22.0
ypy-websocket==0.8.4
zipp==3.23.0
zstandard==0.25.0
